@@ -4,21 +4,22 @@
   "private": true,
   "dependencies": {
     "@ant-design/icons": "^4.7.0",
-    "@reduxjs/toolkit": "^1.7.2",
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.3",
+    "@reduxjs/toolkit": "^1.8.2",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.3.0",
     "@testing-library/user-event": "^13.5.0",
-    "antd": "^4.18.8",
-    "axios": "^0.26.0",
-    "chart.js": "^3.7.1",
-    "html-react-parser": "^1.4.8",
+    "antd": "^4.20.7",
+    "axios": "^0.27.2",
+    "chart.js": "^3.8.0",
+    "html-react-parser": "^1.4.14",
     "millify": "^4.0.0",
-    "moment": "^2.29.1",
-    "react": "^17.0.2",
-    "react-chartjs-2": "^4.0.1",
-    "react-dom": "^17.0.2",
-    "react-redux": "^7.2.6",
-    "react-scripts": "5.0.0",
+    "moment": "^2.29.3",
+    "react": "^18.1.0",
+    "react-chartjs-2": "^4.1.0",
+    "react-dom": "^18.1.0",
+    "react-redux": "^8.0.2",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
