@@ -1,23 +1,29 @@
-import logo from './logo.svg';
-import './App.css';
-
+// import "./App.css";
+import {
+  Navbar,
+  Sidebar,
+  Home,
+  Cryptocurrencies,
+  Exchanges,
+  News,
+} from "./components";
+import "./App.css";
+import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 function App() {
   return (
     <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
+      <Router>
+        <Navbar />
+        <div className="main-wrapper">
+          <Sidebar />
+          <Routes>
+            <Route path="/home" element={<Home />} />
+            <Route path="/cryptocurrencies" element={<Cryptocurrencies />} />
+            <Route path="/exchanges" element={<Exchanges />} />
+            <Route path="/news" element={<News />} />
+          </Routes>
+        </div>
+      </Router>
     </div>
   );
 }
