@@ -1,20 +1,11 @@
-import React from 'react'
-import { Button,Menu,Typography,Avatar } from 'antd';
-import { Link } from 'react-router-dom';
-import { HomeOutlined,MoneyCollectOutlined,BulbOutlined,FundOutlined,MenuOutlined } from '@ant-design/icons';
-import icon from '../images/crptoimg.jpg';
+import React from "react";
+import "../styles/navbar.css";
 const Navbar = () => {
-    return (
-        <div className='nav-container'>
-            <div className='logo-container'>
-                <Avatar src={icon} size='large'/>
-                <Typography.Title level = {2} className ='logo'>
-                    <Link to='/'>CryptApp</Link>
-                </Typography.Title>
-            </div>
-            
-        </div>
-    )
-}
+  return (
+    <>
+      <div className="navbar">fsdfdsaf</div>
+    </>
+  );
+};
 
-export default Navbar
+export default Navbar;
