@@ -1,1 +1,7 @@
-export {default as Navbar} from './Navbar'
+export { default as Navbar } from "./Navbar";
+export { default as HoriNavbar } from "./HoriNavbar";
+export { default as Sidebar } from "./Sidebar";
+export { default as Home } from "./Home";
+export { default as Cryptocurrencies } from "./Cryptocurrencies";
+export { default as Exchanges } from "./Exchanges";
+export { default as News } from "./News";
